CLN: Group shared models into fmu/datamodels/common directory
