+++ src/fmu/datamodels/common/__init__.py
@@ -1,0 +1,38 @@
+from .access import Access, Asset, Ssdl, SsdlAccess
+from .masterdata import (
+    CoordinateSystem,
+    CountryItem,
+    DiscoveryItem,
+    FieldItem,
+    Masterdata,
+    Smda,
+    StratigraphicColumn,
+)
+from .tracklog import (
+    OperatingSystem,
+    SystemInformation,
+    Tracklog,
+    TracklogEvent,
+    User,
+    Version,
+)
+
+__all__ = [
+    "Access",
+    "SsdlAccess",
+    "Asset",
+    "Ssdl",
+    "Masterdata",
+    "Smda",
+    "StratigraphicColumn",
+    "CoordinateSystem",
+    "CountryItem",
+    "FieldItem",
+    "DiscoveryItem",
+    "Tracklog",
+    "OperatingSystem",
+    "SystemInformation",
+    "TracklogEvent",
+    "User",
+    "Version",
+]
